+++ test/instrumentation/modules/aws-sdk/fixtures/dynamodb.js
