+++ test/instrumentation/modules/aws-sdk/fixtures/sqs.js
